+++ packages/dive_obslib/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added example to dive_obslib.
+
 ## [0.2.0]
 
 - Changed method `outputGetState` to `streamOutputGetState`.